Replaced classes with nested functions
- for classes in the view of the UI, we do not
want the boilerplate state initialization
of attributes in the contructor since we think
it will be confusing needing to remember when
an attribute is initialized and when it is
defined. Would rather if the attribute is
initialized and defined at the same time.

- used the original way of nested functions
to allow access of constants to other functions
while not increasing the amount of parameters
passed to functions. But may get a bit confused
from scoping problems of which layer a constant
or function may lie in

- removed constructor in backend to eliminate
having 2 states for attributes (same issue as
classes in the view of the UI)